refactor: 코드구조 변경
