+++ src/main/java/com/softeer/podo/test/model/dto/TestSignUpResponseDto.java
@@ -1,6 +1,6 @@
 package com.softeer.podo.test.model.dto;
 
-import com.softeer.podo.event.model.entity.Role;
+import com.softeer.podo.user.model.entity.Role;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
