+++ src/main/java/com/softeer/podo/user/model/entity/Role.java
@@ -1,4 +1,4 @@
-package com.softeer.podo.event.model.entity;
+package com.softeer.podo.user.model.entity;
 
 public enum Role {
     ROLE_USER, ROLE_ADMIN, ROLE_ANONYMOUS
